optimize some git-diff/upstream behavior
This applies on branches with no upstream; previously we ran
`git diff --quiet @{u}` regardless, and tested the exit code to see if
we needed to fall back to diffing against HEAD. This command is fairly
quick (in the grand scheme of things) if there's no upstream--but not
as fast as just getting this information from libgit2. On my system it
saves ~25ms per query on a nixpkgs branch w/ no upstream.